+++ packages/orq-rc/src/orq_ai_sdk/models/createconversationop.py
@@ -199,7 +199,7 @@ class CreateConversationResponseBody(BaseModel):
     r"""Unix timestamp in milliseconds when the conversation was last modified."""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kgha06pqg25rjv54gsvx26ty"
+        "conv_01kghk7xr1e12540fq96h8rhaw"
     )
     r"""Unique conversation identifier with `conv_` prefix."""