+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KGHA0776G9A3NH58EGGQ3QMA"
+        "01KGHK7Y8BTR0ZPJX88PS20355"
     )
     r"""The unique identifier of the data source"""