+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KGHA0775HQSRDXKQGGNVYZT0"
+        "01KGHK7Y8AK3Q6XAGYYH7DRQBH"
     )
     r"""The unique identifier of the data source"""