+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KGHA0779T1XMX2DEDN5RKSBC"
+        "01KGHK7Y8F71AY0XHJ7V0CCSX5"
     )
     r"""The unique identifier of the data source"""