+++ packages/orq-rc/RELEASES.md
@@ -5209,3 +5209,13 @@ Based on:
 - [python v4.3.0-rc.19] packages/orq-rc
 ### Releases
 - [PyPI v4.3.0-rc.19] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.19 - packages/orq-rc
+
+## 2026-02-03 11:14:45
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.700.2 (2.801.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.3.0-rc.20] packages/orq-rc
+### Releases
+- [PyPI v4.3.0-rc.20] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.20 - packages/orq-rc